Converted to detector
