+++ aderyn_core/src/audit/auditor.rs
@@ -6,7 +6,6 @@ use crate::{
     audit::{
         attack_surface::AttackSurfaceDetector,
         public_functions_no_sender::PublicFunctionsNoSenderChecksDetector,
-        send_ether_no_checks::SendEtherWithoutMsgSenderChecksDetector,
     },
     context::workspace_context::WorkspaceContext,
 };
@@ -15,7 +14,6 @@ pub fn get_auditor_detectors() -> Vec<Box<dyn AuditorDetector>> {
     vec![
         Box::<AttackSurfaceDetector>::default(),
         Box::<PublicFunctionsNoSenderChecksDetector>::default(),
-        Box::<SendEtherWithoutMsgSenderChecksDetector>::default(),
     ]
 }
