+++ aderyn_core/src/audit/mod.rs
@@ -7,4 +7,3 @@ pub mod investigators;
 // AuditorDetectors
 pub mod attack_surface;
 pub mod public_functions_no_sender;
-pub mod send_ether_no_checks;
