+++ aderyn_core/src/detect/high/mod.rs
@@ -11,6 +11,7 @@ pub(crate) mod multiple_constructors;
 pub(crate) mod nested_struct_in_mapping;
 pub(crate) mod reused_contract_name;
 pub(crate) mod selfdestruct;
+pub(crate) mod send_ether_no_checks;
 pub(crate) mod state_variable_shadowing;
 pub(crate) mod storage_array_edit_with_memory;
 pub(crate) mod unprotected_init_function;
@@ -30,6 +31,7 @@ pub use multiple_constructors::MultipleConstructorsDetector;
 pub use nested_struct_in_mapping::NestedStructInMappingDetector;
 pub use reused_contract_name::ReusedContractNameDetector;
 pub use selfdestruct::SelfdestructIdentifierDetector;
+pub use send_ether_no_checks::SendEtherNoChecksDetector;
 pub use state_variable_shadowing::StateVariableShadowingDetector;
 pub use storage_array_edit_with_memory::StorageArrayEditWithMemoryDetector;
 pub use unprotected_init_function::UnprotectedInitializerDetector;
